% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_awesomecv.R
\name{sr_awesomecv}
\alias{sr_awesomecv}
\title{Awesome CV template}
\usage{
sr_awesomecv(
  ...,
  latex_engine = "xelatex",
  page_total = FALSE,
  show_footer = TRUE
)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[vitae]{cv_document}}.}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", "xelatex" and "tectonic".}

\item{page_total}{If TRUE, the total number of pages is shown in the footer.}

\item{show_footer}{If TRUE, a footer showing your name, document name, and page number.}
}
\value{
An R Markdown output format object.
}
\description{
Awesome CV is LaTeX template for a CV or Résumé inspired by
\href{https://www.overleaf.com/latex/templates/friggeri-cv-template/hmnchbfmjgqh}{Fancy CV}:
https://github.com/posquit0/Awesome-CV
}
